product class
The code defines a `Product` class that represents a table in a database. It has two columns: `product_id` and `date_created`. The `product_id` column is the primary key of the table, and the `date_created` column is a timestamp that defaults to the current UTC time. The `__repr__` method is overridden to provide a string representation of a `Product` object.