ProductMovement class
The code defines a `ProductMovement` class that represents a table in a database. It has columns for `movement_id`, `product_id`, `qty`, `from_location`, `to_location`, and `movement_time`. 

The class also defines relationships with the `Product` and `Location` tables using foreign keys. 

The `__repr__` method is overridden to provide a string representation of a `ProductMovement` object.