Flask route handler
The code defines a Flask route handler for the root URL ("/"). 

If the request method is POST and the form field "product_name" is present, a new `Product` object is created with the provided name and added to the database. If the operation is successful, the user is redirected to the root URL. Otherwise, an error message is returned.

If the request method is POST and the form field "location_name" is present, a new `Location` object is created with the provided name and added to the database. If the operation is successful, the user is redirected to the root URL. Otherwise, an error message is returned.

If neither of the above conditions are met, the code retrieves all `Product` and `Location` objects from the database and renders the "index.html" template, passing the retrieved objects as variables.