route '/locations/'
The code defines a route `/locations/` in a Flask application. When the route is accessed via a POST request and the form data contains a field called `location_name`, a new `Location` object is created and added to the database. If the addition is successful, the user is redirected to `/locations/`. If there is an error, a list of all locations is retrieved from the database and rendered in a template called `locations.html`. If the route is accessed via a GET request, the list of all locations is retrieved from the database and rendered in the `locations.html` template.