viewProduct
The code defines a route `/products/` in a Flask application. When a request is made to this route, the `viewProduct` function is executed. 

If the request method is POST and the form data contains a field called `product_name`, a new `Product` object is created with the `product_name` as the `product_id`. It is then added to the database using SQLAlchemy and a redirect to `/products/` is returned.

If an error occurs during the database operation, all existing products are fetched from the database and an error message is returned.

If the request method is not POST or the form data does not contain `product_name`, all existing products are fetched from the database and rendered in a template called "products.html".