`/update-movement/<int:id>`
The code defines a route `/update-movement/<int:id>` that handles both GET and POST requests. 
When a GET request is made, it retrieves a `ProductMovement` object with the given `id` from the database, 
as well as all `Product` and `Location` objects from the database. 
It then renders a template called "update-movement.html" with the retrieved objects as arguments.

When a POST request is made, it updates the properties of the `ProductMovement` object with the values 
from the submitted form data. It then attempts to commit the changes to the database and redirects to 
"/movements/" if successful. If an error occurs during the commit, it returns an error message.