route '/product-balance/'
The code defines a route `/product-balance/` in a Flask application. When the route is accessed via a POST or GET request, the `productBalanceReport` function is executed. 

Inside the function, a query is made to retrieve data from the `ProductMovement` table, joining it with the `Product` table. The retrieved data is then processed and stored in a nested dictionary called `balancedDict`. 

The `balancedDict` dictionary is used to calculate and store the balance of product quantities between different locations. The final result is passed to a template called "product-balance.html" for rendering.