route `/movements/get-from-locations/`
The code defines a route `/movements/get-from-locations/` that handles a POST request. 
The function `getLocations` retrieves data from the request form, queries a database table called `ProductMovement`, 
and constructs a dictionary `locationDict` that maps each `to_location` to the sum of the corresponding `qty` values. 
Finally, the function returns the `locationDict` dictionary.